More proper package structure
